chat-service-python/requirements.txt
Django==5.1
openai==1.43.0
python-dotenv==1.0.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
azure-monitor-opentelemetry-exporter==1.0.0b28
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-django==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
# opentelemetry-instrumentation-openai==0.48b0 - install from https://github.com/lmolkova/opentelemetry-python-contrib/blob/openai-events/instrumentation/opentelemetry-instrumentation-opena

